requirements.txt
bleach==3.3.1
canonicalwebteam.flask-base==0.9.3
canonicalwebteam.http==1.0.3
canonicalwebteam.image-template==1.3.1
canonicalwebteam.templatefinder==1.0.0
markdown==3.3.6
python-slugify==4.0.1
vcrpy-unittest==0.1.7
cryptography==36.0.1
